refactor: 게임 진행 및 결과 출력 부분을 새로운 클래스로 분리
